+++ src/api/index.ts
@@ -1,3 +1,4 @@
 import { loginWithGoogle } from './auth';
+import axios from './api';
 
-export { loginWithGoogle };
+export { loginWithGoogle, axios };
