--- src/components/App/App.tsx
@@ -1,24 +1,0 @@
-import Header from '../Common/Header';
-
-const App = () => {
-  return (
-    <div className='font-NotoSans'>
-      <Header />
-      <div className='container'>
-        {/* <Routes> */}
-        {/* <Route path='/' element={<ListContainer />} />
-        <Route path='/info' element={<UserContainer />} />
-        <Route path='/sentence' element={<UserContainer />} />
-        <Route path='/likes' element={<UserContainer />} />
-        <Route path='/user-detail/:id' element={<DetailContainer />} />
-        <Route path='/book-detail/:id' element={<DetailContainer />} />
-        <Route path='/search' element={<ListKeywordSearch />} />
-        <Route path='/add' element={<BookContainer />} />
-        <Route path='/modify/:id' element={<BookContainer />} /> */}
-        {/* </Routes> */}
-      </div>
-    </div>
-  );
-};
-
-export default App;
