--- src/components/Common/Button/index.stories.ts
@@ -1,11 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-import Button from './';
-const meta = {
-  title: 'common/Button',
-  component: Button,
-  tags: ['autodocs'],
-} satisfies Meta<typeof Button>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-export const Default: Story = {};
