--- src/components/Common/MaxWidthWrapper/index.tsx
@@ -1,23 +1,0 @@
-import { ReactNode } from 'react';
-import { cn } from '../../../utils';
-
-const MaxWidthWrapper = ({
-  className,
-  children,
-}: {
-  className?: string;
-  children: ReactNode;
-}) => {
-  return (
-    <div
-      className={cn(
-        'mx-auto w-full max-w-screen-xl px-2.5 md:px-20',
-        className
-      )}
-    >
-      {children}
-    </div>
-  );
-};
-
-export default MaxWidthWrapper;
