--- src/components/Common/index.tsx
@@ -1,4 +1,0 @@
-import Logo from './Logo';
-import MaxWidthWrapper from './MaxWidthWrapper';
-
-export { Logo, MaxWidthWrapper };
