+++ src/components/index.tsx
@@ -1,0 +1,4 @@
+import Logo from './Logo';
+import MaxWidthWrapper from './MaxWidthWrapper';
+
+export { Logo, MaxWidthWrapper };
