--- src/store/actions/actionTypes.js
@@ -1,42 +1,0 @@
-// common
-export const TOGGLE_MENU = 'TOGGLE_MENU';
-export const TOGGLE_POPUP = 'TOGGLE_POPUP';
-export const TOGGLE_SORT = 'TOGGLE_SORT';
-export const CHANGE_LOADGIN_STATUS = 'CHANGE_LOADGIN_STATUS';
-
-// user
-export const LOGIN_WITH_FIREBASE = 'LOGIN_WITH_FIREBASE';
-export const CHANGE_LOGIN_STATUS  = 'CHANGE_LOGIN_STATUS';
-export const GET_FIREBASE_USER_DATA = 'GET_FIREBASE_USER_DATA';
-export const SET_USER_INFO = 'SET_USER_INFO';
-export const SET_USER_ID = 'SET_USER_ID';
-export const SET_USER_LIKES_LIST = 'SET_USER_LIKES_LIST';
-export const CHANGE_NAME_INPUT = 'CHANGE_NAME_INPUT';
-export const CHANGE_NAME = 'CHANGE_NAME';
-export const CHANGE_USER_PICTURE = 'CHANGE_USER_PICTURE';
-
-// list
-export const SET_SENTENCE_LIST = 'SET_SENTENCE_LIST';
-export const SHOW_MORE_SENTENCE_BODY = 'SHOW_MORE_SENTENCE_BODY';
-export const CLEAR_LIST_ITEM = 'CLEAR_LIST_ITEM';
-export const CHAHNGE_LIST_ITEM = 'CHAHNGE_LIST_ITEM';
-export const TOGGLE_MODIFY_BUTTON = 'TOGGLE_MODIFIED_BUTTON';
-
-// book
-export const RESET_BOOK_STATE = 'RESET_BOOK_STATE';
-export const SHOW_INPUT_ALERT = 'SHOW_INPUT_ALERT';
-export const INPUT_BOOK_SEARCH = 'INPUT_BOOK_SEARCH';
-export const SET_BOOK_SEARCH_RESULT = 'SET_BOOK_SEARCH_RESULT';
-export const SELECT_SEARCHED_BOOK = 'SELECT_SEARCHED_BOOK';
-export const SHOW_BOOKINFO_INPUT = 'SHOW_BOOKINFO_INPUT';
-export const CHANGE_BOOKINFO_INPUT = 'CHANGE_BOOKINFO_INPUT';
-export const SUBMIT_BOOKINFO_INPUT = 'SUBMIT_BOOKINFO_INPUT';
-export const CHANGE_SENTENCE_TEXTAREA = 'CHANGE_SENTENCE_TEXTAREA';
-export const SUBMIT_SENTENCE = 'SUBMIT_SENTENCE';
-export const SET_MODIFY_STATUS = 'SET_MODIFY_STATUS';
-
-//detail 
-export const SET_DETAIL_LIST = 'SET_DETAIL_LIST';
-export const SET_SELECTED_BOOK_INFO = 'SET_SELECTED_BOOK_INFO';
-export const SET_SELECTED_USER_INFO = 'SET_SELECTED_USER_INFO';
-export const CHANGE_DETAIL_TAB = 'CHANGE_DETAIL_TAB';
