--- src/store/actions/index.js
@@ -1,6 +1,0 @@
-export * from './commonAction';
-export * from './listAction';
-export * from './userAction';
-export * from './bookAction';
-export * from './detailAction';
-
