--- src/store/actions/listAction.js
@@ -1,141 +1,0 @@
-import * as types from './actionTypes';
-import { firestore } from '../../modules/firebaseConfig';
-import { getDetailListFromDB } from './detailAction';
-import { getUserLikesListDB, getUserSentenceListDB } from './userAction';
-import { changeLoadingStatus } from './commonAction';
-
-
-// 리스트==================================
-
-// DB에서 가져온 리스트를 출력형식에 맞게 정리
-export const getList = ({ docs, userId, orderBy }) => dispatch => {
-  const list = docs.map(doc => {
-    const { body } = doc.data();
-    // 본문 길이 체크
-    const bodyLengthCheck =body.length > 200
-      ? { printBody: body.slice(0, 200) + '...', showMoreButton: true }
-      : { printBody: body, showMoreButton: false };
-      
-    return Object.assign({}, { id: doc.id }, doc.data(), { time: doc.data().updateDate.toDate() }, bodyLengthCheck, { showMore: false });
-  });
-  const lastItem = list.length &&  list[list.length - 1].id;
-  // 책에 등록된 문장 리스트에서 내가 등록한 문장을 구분하기 위해서 userId로 필터링한다. 
-  if (userId) {
-    const userList = list.filter(item => item.userInfo.id.indexOf(userId) > -1);
-    dispatch(setSentenceList({ list, lastItem, userList, orderBy }));
-  } else {
-    const userList = [];
-    dispatch(setSentenceList({ list, lastItem, userList, orderBy }));
-  }
-  dispatch(changeLoadingStatus(false));
-}
-// DB에서 리스트 가져오기
-export const getSentenceListFromDB = (orderBy, startItem) => dispatch => {
-  const sentenceRef = firestore.collection("sentences");
-  // 처음으로 List를 가져올 때 
-  if (!startItem) {
-    dispatch(clearListItem());
-    sentenceRef.orderBy(orderBy, "desc").limit(5).get().then(snapshot => {
-      // 쿼리 결과가 있을 때만 리스트를 출력한다
-      if (snapshot.docs.length > 0) {
-        dispatch(getList({ docs: snapshot.docs, orderBy }));
-      } else {
-        // 출력할 리스트가 없다면 스피너를 숨긴다.
-        dispatch(changeLoadingStatus(false));
-      }
-    });
-  } else {
-    // 추가로 데이터 호출
-    sentenceRef.doc(startItem).get().then(snapshot => {
-      sentenceRef.orderBy(orderBy, "desc").startAfter(snapshot).limit(5).get().then(snapshot => {
-        // 쿼리 결과가 있을 때만 리스트를 출력한다
-        if(snapshot.docs.length > 0) {
-          dispatch(getList({ docs: snapshot.docs, orderBy }));
-        } else {
-          // 추가로 출력할 리스트가 없다면 스피너를 숨긴다.
-          dispatch(changeLoadingStatus(false));
-        }
-      });
-    })
-  }
-}
-
-export const setSentenceList = ({list, lastItem, orderBy, userList}) => ({
-  type: types.SET_SENTENCE_LIST,
-  list,
-  lastItem,
-  orderBy,
-  userList
-})
-
-// 문장 더보기
-export const showMoreSentenceBody = (index) => {
-  return {
-    type: types.SHOW_MORE_SENTENCE_BODY,
-    index
-  }
-}
-
-// 리스트 비우기 > 최신순,인기순 정렬시 이전 List에 추가되어버리기 떄문에
-export const clearListItem = () => {
-  return {
-    type: types.CLEAR_LIST_ITEM
-  }
-}
-
-// 좋아요 클릭 
-export const handleLikeClick = (index, id, likes, userId) => dispatch => {
-  // 해당 문장의 user정보를 비교하여 좋아요를 추가해야하는지 취소해야하는지 판단
-  const sentenceRef = firestore.collection('sentences').doc(id);
-  sentenceRef.get().then(snapshot => {
-    const data = snapshot.data();
-    const likeUser = data.likeUser, userIndex = likeUser.indexOf(userId);
-    let updateLikeUserData = [];
-    if (userIndex < 0) {
-      likes++;
-      updateLikeUserData = [...likeUser, userId];
-    } else {
-      likes--;
-      updateLikeUserData = likeUser.filter(user => user !== userId);
-    }
-    sentenceRef.update({ likeUser: updateLikeUserData, likes: likes });
-    dispatch(changeListItem(index, 'likes', likes));
-    dispatch(changeListItem(index, 'likeUser', updateLikeUserData));
-  });
-}
-
-// state의 단일 요소 수정
-export const changeListItem = (index, key, value) => {
-  return {
-    type: types.CHAHNGE_LIST_ITEM,
-    index,
-    key,
-    value
-  }
-}
-
-// 글 수정, 삭제 버튼 토글
-export const toggleModifyButton = (bool, id ) =>{
-  return {
-    type: types.TOGGLE_MODIFY_BUTTON,
-    bool, 
-    id
-  }
-}
-
-// 글 삭제
-export const deleteListItem = ({sentenceId, path, getListDB}) => dispatch => {
-  firestore.collection('sentences').doc(sentenceId).delete();
-  dispatch(clearListItem());
-  
-  // 글을 삭제한 곳에 따라 다른 DB 요청
-  if(path === '/') {
-    dispatch(getSentenceListFromDB('updateDate'));
-  } else if (path.indexOf('detail') > -1) {
-    dispatch(getDetailListFromDB(getListDB));
-  } else if (path.indexOf('likes') > -1) {
-    dispatch(getUserLikesListDB(getListDB))
-  } else if (path.indexOf('sentence') > -1) {
-    dispatch(getUserSentenceListDB(getListDB))
-  }
-}
