--- src/store/reducers/commonReducer.js
@@ -1,28 +1,0 @@
-
-import { Map } from 'immutable';
-import * as actions from '../actions/actionTypes';
-
-// common 
-const commonInitialState = Map({
-  isMenuOpen: false,
-  isSortOpen: false,
-  showPopup: false,
-  popupMsg: '',
-  target: '',
-  isLoading: false
-});
-
-export const common = (state = commonInitialState, action) => {
-  switch (action.type) {
-    case actions.TOGGLE_MENU:
-      return state.set('isMenuOpen', !state.get('isMenuOpen'));
-    case actions.TOGGLE_SORT: 
-      return state.set('isSortOpen', !state.get('isSortOpen'));
-    case actions.TOGGLE_POPUP:
-      return state.merge({ showPopup: !state.get('showPopup'), popupMsg: action.msg });
-    case actions.CHANGE_LOADGIN_STATUS:
-      return state.set('isLoading', action.bool);
-    default:
-      return state;
-  }
-}
