--- src/styles/components/Intro.scss
@@ -1,95 +1,0 @@
-@import '../modules/_all.scss'; // 로그인 전 인트로
-.intro {
-  height: 100vh;
-  font-size: 35px;
-  color: #fff;
-  background: #333;
-  overflow: hidden;
-  .intro-slide {
-    display: flex;
-    width: 500%;
-    li.slide-item {
-      padding: 100px 20px 20px;
-      width: 100%;
-      text-align: center;
-      animation: slide-animation 20s ease-in-out infinite;
-      .from {
-        margin-top: 10px;
-        font-size: 30px;
-      }
-    }
-  }
-  .btn-login {
-    position: fixed;
-    left: 50%;
-    bottom: 20%;
-    width: 250px;
-    height: 40px;
-    transform: translateX(-50%);
-    font-family: $font-default;
-    font-size: 20px;
-    button {
-      width: 100%;
-      height: 100%;
-      border: none;
-      border-radius: 5px;
-      background: #dd4b39;
-      box-shadow: 3px 3px#b63524;
-       color: #fff;
-      cursor: pointer;
-      .google-icon {
-        margin-right: 20px;
-        color: #fff;
-      }
-    }
-  }
-  .desktop-alert {
-    display: none;
-  }
-  @keyframes slide-animation {
-    0% {
-      transform: translateX(100%);
-    }
-    1% {
-      transform: translateX(0);
-    }
-    18% {
-      transform: translateX(0);
-    }
-    20% {
-      transform: translateX(-100%);
-    }
-    38% {
-      transform: translateX(-100%);
-    }
-    40% {
-      transform: translateX(-200%);
-    }
-    58% {
-      transform: translateX(-200%);
-    }
-    60% {
-      transform: translateX(-300%);
-    }
-    78% {
-      transform: translateX(-300%);
-    }
-    80% {
-      transform: translateX(-400%);
-    }
-    98% {
-      transform: translateX(-400%);
-    }
-    100% {
-      transform: translateX(-500%);
-    }
-  }
-}
-
-@include breakpoint(large) {
-  .intro {
-    .intro-slide {
-      padding-top: 150px;
-    }
-  }
-}
