--- src/styles/components/List.scss
@@ -1,18 +1,0 @@
-@import '../modules/_all.scss';
-.main-list {
-  .add-button {
-    position: fixed;
-    bottom: 20px;
-    right: 20px;
-    width: 70px;
-    height: 70px;
-    font-size: 70px;
-    font-weight: bold;
-    text-align: center;
-    line-height: 70px;
-    background: $base-color;
-    box-shadow: 2px 2px $base-color-dark;
-    border-radius: 50%;
-    z-index: 500;
-  }
-}
