--- src/styles/components/ListItem.scss
@@ -1,162 +1,0 @@
-@import '../modules/_all.scss';
-
- .list-sort {
-   position: relative;
-   text-align: right;
-   button {
-     padding-left: 10px;
-     width: 100px;
-     height: 30px;
-     background: #fff;
-     border: 1px solid#aaa;
-     text-align: left;
-     cursor: pointer;
-     .arrow-icon {
-       margin-left: 30px;
-       color: #555;
-       font-size: 20px;
-       line-height: 30px;
-     }
-   }
-   .select-list {
-     position: absolute;
-     right: 0;
-     z-index: 300;
-     button {
-       border-top: none;
-     }
-   }
- }
- .sentence-list {
-   padding-top: 20px;
-   .list-item {
-     position: relative;
-     padding: 30px 20px;
-     margin-bottom: 50px;
-     min-height: 400px;
-     border: 1px solid #333;
-     border-radius: 5px;
-     .info {
-       position: relative;
-       margin-bottom: 10px;
-       text-align: center;
-      .list-modify{
-        position: absolute;
-        top: -15px;
-        right: 0;
-        text-align: right;
-        button.toggle-button, .select-item button {
-          background-color: #fff;
-          border: none;
-          cursor: pointer;
-        }
-        .select-list {
-          position: absolute;
-          right: 0px;
-          z-index: 300;
-          li.select-item {
-            width: 60px;
-            border: 1px solid #aaa;
-            text-align: center;
-            background-color: #fff;
-            &:first-child {
-              border-bottom: 0;
-            }
-  
-          }
-        }
-      }
-      img.book {
-        margin-bottom: 20px;
-        box-shadow: 4px 4px #ccc;
-      }
-      p.book.title {
-        margin-bottom: 10px;
-        padding-top: 5px;
-        font-size: 25px;
-        font-weight: bold;
-      }
-      .user.date {
-        margin-top: 10px;
-        display: flex;
-        flex-wrap: wrap;
-        justify-content: space-between;
-        a.user {
-            display: flex;
-          img.user {
-            margin-right: 10px;
-            width: 30px;
-            height: 30px;
-            border-radius: 50%;
-          }
-           .user.name {
-             margin-right: 10px;
-             line-height: 30px;
-           }
-         }
-         .update-date {
-           line-height: 30px;
-           font-size: 14px;
-         }
-       }
-       &::after {
-         content: '';
-         display: block;
-         clear: both;
-       }
-     }
- }
-
-   .sentence {
-     text-align: right;
-
-     .body {
-       margin: 20px 0;
-       font-size: 17px;
-       line-height: 1.8;
-       text-align: left;
-     }
-
-     .more {
-       background: #fff;
-       border: none;
-       cursor: pointer;
-     }
-   }
-
-   .likes {
-     position: absolute;
-     left: 15px;
-     bottom: 15px;
-     background: #fff;
-     border: none;
-     color: red;
-     cursor: pointer;
-
-     span {
-       color: $font-color;
-     }
-   }
- }
-
- @include breakpoint(large) {
-    .sentence-list {
-      display: flex;
-      flex-wrap: wrap;
-      .list-item {
-        margin: 15px;
-        width: 29%;
-      }
-    }
-  }
-
- @include breakpoint(medium) {
-    .sentence-list {
-      display: flex;
-      flex-wrap: wrap;
-      .list-item {
-        margin: 15px;
-        width: 45%;
-      }
-    }
- }
