--- src/styles/components/Nav.scss
@@ -1,111 +1,0 @@
-@import '../modules/_all.scss';
-.top-bar {
-  height: 50px;
-  .default {
-    position: fixed;
-    top: 0;
-    left: 0;  
-    display: flex;
-    justify-content: center;
-    padding: 10px;
-    width: 100%;
-    font-size: 35px;
-    background: $base-color;
-    z-index: 3000;
-    button {
-      position: absolute;
-      left: 5px;
-      height: 30px;
-      border: none;
-      background: $base-color;
-      font-size: 30px;
-      cursor: pointer;
-    }
-    a {
-      line-height: 30px;
-      font-weight: bold;
-    }
-  }
-  .nav-menu {
-    position: absolute;
-    width: 80vw;
-    height: calc(100vh - 50px);
-    top: 50px;
-    left: -110%;
-    transition: 0.5s ease-out;
-    background: #fff;
-    box-shadow: 3px 0px 5px rgba(0, 0, 0, 0.9);
-    z-index: 2500;
-    &.is-open {
-      left: 0;
-    }
-    .user-info{
-      padding: 40px 10px 10px 10px;
-      text-align: center;
-      img {
-        width: 150px;
-        height: 150px;
-        border-radius: 50%;
-        text-align: center;
-      }
-      .name {
-        margin-top: 15px;
-        font-weight: bold;
-        font-size: 30px;
-      }
-    }
-    ul {
-      padding-top: 20px;
-      text-align: center;
-      li {
-        margin-top: 20px;
-        font-size: 20px;
-      }
-    }
-    button.logout {
-      position: relative;
-      left: 65%;
-      margin-top: 30px;
-      width: 100px;
-      height: 30px;
-      color: #fff;
-      font-size: 20px;
-      background: $base-color;
-      border: none;
-      border-radius: 5px;
-      box-shadow: 1px 1px $base-color-dark;
-      cursor: pointer;
-    } 
-  }
-  .nav-background {
-    position: absolute;
-    left: -110%;
-    top: 0;
-    width: 100vw;
-    height: 100vh;
-    background: rgba(0, 0, 0, 0.8);
-    z-index: 2000;
-    cursor: pointer;
-    &.is-open {
-      left: 0;
-    }
-  }
-}
-
-@include breakpoint(large) {
-  .top-bar {
-    .nav-menu {
-      width: 30vw;
-      left: -50%;
-    }
-  }
-}
-
-@include breakpoint(medium) {
-  .top-bar {
-    .nav-menu {
-      width: 50vw;
-      left: -60%;
-    }
-  }
-}
