--- src/styles/components/User.scss
@@ -1,119 +1,0 @@
-@import '../modules/_all.scss';
-
-.user-sentence, .user-likes, .user-info {
-  .page-title{
-    margin-bottom: 10px;
-    font-size: 30px;
-    font-weight: bold;
-  }
-  .count {
-    font-size: 25px;
-  }
-}
-
-.user-info {
-  padding: 30px;
-  text-align: center;
-  .page-title {
-    text-align: left;
-  }
-  .info-table {
-    width: 100%;
-    margin-left: 50%;
-    transform: translateX(-50%);
-    margin-top: 30px;
-    font-size: 20px;
-    text-align: left;
-    td {
-      padding: 10px 5px;
-      &.label {
-        width: 25%;
-        text-align: right;
-      }
-      &.picture {
-        text-align: center;
-        padding-bottom: 40px;
-        label {
-          position: relative;
-          img {
-            width: 150px;
-            height: 150px;
-            border-radius: 50%;
-          }
-          &:hover {
-            &::after {
-              content: "사진 변경";
-              position: absolute;
-              left: 0;
-              display: inline-block;
-              width: 150px;
-              height: 150px;
-              color: #fff;
-              line-height: 150px;
-              text-align: center;
-              background: rgba(0, 0, 0, 0.5);
-              border-radius: 50%;
-              cursor: pointer;
-            }
-          }
-        }
-        #file-input{
-          position: absolute !important;
-          height: 1px;
-          width: 1px;
-          overflow: hidden;
-          clip: rect(1px 1px 1px 1px);
-          /* IE6, IE7 */
-          clip: rect(1px, 1px, 1px, 1px);
-        }
-      }
-      input {
-        margin-right: 10px;
-        padding: 3px;
-        width: 100%;
-        border: none;
-        border-bottom: 2px solid $base-color;
-      }
-      button {
-        width: 60px;
-        font-size: 18px;
-      }
-    }
-    &:after{
-      content: '';
-      display: block;
-      clear: both;
-    }
-  }
-  button {
-    margin-top: 10px;
-    width: 100px;
-    height: 30px;
-    background-color: $base-color;
-    border: 1px solid $base-color-dark;
-    box-shadow: 1px 1px $base-color-dark;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  .user-delete {
-    position: relative;
-    right: -45%;
-    margin-top: 40px;
-  }
-}
-
- @include breakpoint(large) {
-   .user-info {
-    .info-table {
-      width: 500px;
-    }
-   }
- }
-
- @include breakpoint(medium) {
-   .user-info {
-     .info-table {
-       width: 500px;
-     }
-   }
- }
