--- src/styles/components/index.scss
@@ -1,30 +1,0 @@
-@import '../modules/_all.scss';
-
-body {
-  font-family: $font-default;
-}
-
-body * {
-  font-family: inherit;
-  box-sizing: border-box;
-}
-
-body .intro {
-  font-family: $font-intro;
-}
-
-ul {
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-
-h1, h2, h3, h4, h5, h6, p {
-  margin: 0;
-   word-break: keep-all;
-}
-
-a {
-  text-decoration: none;
-  color: inherit;
-}
