--- src/styles/modules/_all.scss
@@ -1,5 +1,0 @@
-
-@import './_normalize.scss';
-@import './_mixin.scss';
-@import './_color.scss';
-@import './_font.scss';
