--- src/styles/modules/_color.scss
@@ -1,4 +1,0 @@
-
-$base-color: #F26A8D;
-$base-color-dark: #861a10;
-$font-color: #333;
