--- src/styles/modules/_font.scss
@@ -1,13 +1,0 @@
-@font-face {
-  font-family: 'KCCEunYeong';
-  src: url('../../assets/KCCEunYeongche.ttf');
-}
-
-@font-face {
-  font-family: 'NanumBarunpen';
-  src: url('../../assets/NanumBarunpenR.ttf');
-}
-
-
-$font-default: NanumBarunpen, Helvetica, Arial, sans-serif;
-$font-intro: KCCEunYeong, Georgia, Myeongjo, serif;
