--- src/styles/modules/_mixin.scss
@@ -1,22 +1,0 @@
-/*********************
-breakpoint
-*********************/
-@mixin breakpoint($point) {
- @if $point==large {
-    @media (min-width: 1024px) {
-      @content ;
-    }
-  }
-  @if $point==medium {
-    @media (min-width: 769px) and (max-width: 1023px) {
-      @content ;
-    }
-  }
-
-  @else if $point==small {
-    @media (max-width: 768px) {
-      @content ;
-    }
-
-  }
-}
